tests: Add a partial test for gtk_doc check: kwarg
check: true is needed for an actual regression test, but that requires
a pedantically correct gtk-doc configuration, which I attempted to do
and failed. So let's just put check: false so we get *some* coverage:
just that typed_kwargs accepts the argument.

Related to https://github.com/mesonbuild/meson/pull/9807