compilers: Convert b_sanitize to a freeform array option
There are a lot of sanitizers these days, and trying to keep up with
them would be entirely too much work. Instead we'll do a compile check
to see if the option is valid. An array makes more sense than a string,
since users could just write `address,undefined` or they could write
`undefined,address`. With the array we get to handle that ourselves and
users writing code know that it's not safe to simply do a check like
```meson
if get_option('b_sanitize') == 'address,undefined'
  ...
endif
```

instead they need to do something like
```meson
opt = get_option('b_sanitize')
if opt.contains('address') and opt.contains('undefined')
   ...
endif
```

To ensure backwards compatibility, a special check has been added to the
get_option() call for `b_sanitize`, which converts it to a string if the
compatibility is < 0.63, with a warning that the behavior will change in
the future.

Fixes #8283
Fixes #7761
Fixes #5154
Fixes #1582