mtest: replace the status line with a multi line one
Rotating the names of run tests on a single line can make it difficult
to read them (or alternatively the rotation becomes very slow). Instead
of displaying one of the running tests at a time on the status line use
multiple status lines to list the statuses of multiple running tests.
Make it so that the longest running tests are shown first and the rest
are shown if there is enough room for them.

The number of tests being run at a given time shown in the form of
$n-$m/$y can be difficult to read (and do math on) when tests are quick
and the numbers change rapidly. When displaying any counts of running
tests, keep the numbers as static as possible to make them easy to read.

Limit the number of tests / status lines to 10 to make things manageable
even with large numbers of tests running in parallel. Let the user know
how many tests are being run in the backround with an additional status
line.