mtest: use a simple status message/line in quiet mode
The current status line(s) is suboptimal when run in quiet mode (it is
not that "quiet").

Implement a simple, "quiet", (updating) status message stating the test
counts / progress in an easy to ready format. As with the main status
line(s) avoid having any rapidly moving parts in it to ensure easy
reading even when running lots of short tests.