interpreter: deprecate using an array for project(license:)
It's ambiguous as to whether the array means AND or OR, and ambiguity in
software licenses is *not* a good thing. Further, SPDX (which we
generally recommend) is an ISO standard and well understood outside of
Meson (including in SBOM software). Deprecate passing and array and
suggest using an SDPX expression instead.

If an older version is targeted, then a warning is generated if
using an array with a length > 1 to use an SPDX expression instead.