minstall: raise explicit errors, by using MesonException
RuntimeError is way too generic when we have an explicit class for
"Meson reports to the user, something went wrong".

Moreover we now tell people that generic exceptions are "Meson bugs and
should be reported", so our failure to do the technically correct thing
and report the right kind of exception means we get haunted by demons of
confusion. Specifically, people complain to us that Meson told them
"there is a bug in Meson" when their install fails due to meson.build or
build environment issues.