pkgconfig: Include libdir in .pc files if referenced
Some packages define variables in their .pc files and those variables
may depend on the value of the libdir variable even if no public or
private libraries are referenced in the .pc file.

Walk the list of variables and check if any of them reference ${libdir}
and if so, make sure to include the libdir definition as well.