test cases: rename shared library('c') avoid libc collision
In a bunch of cases we create a series of sample libraries named "a",
"b", "c" etc.

This breaks on musl. Originally reported with muon via commit
https://git.sr.ht/~lattis/muon/commit/ca5c37171423e9884047a85349e1b236d7449510
and also breaks the testsuite when packaging meson for alpine linux.

libc.so is an existing library which is linked in by default for all the
obvious reasons. You can get away with this on glibc, because that
includes a soversion of "6", but it loads the wrong library on musl.