CI images: don't build on a schedule for forks
It's a bit annoying having to *manually* disable a workflow for forks,
because the scheduled build is a waste of time and also spammy.

On the other hand, having the workflow enabled is in general a good
thing, because it is in fact desirable to test it locally in forks.